--- js/generate-random.js
@@ -1,7 +1,0 @@
-export function randomInteger (min, max) {
-
-  const rand = min + Math.random() * (max + 1 - min);
-  return Math.floor(rand);
-}
-
-
