--- js/get-check-length.js
@@ -1,6 +1,0 @@
-function checkLength(str, maxLength) {
-  return str.length > maxLength;
-
-}
-
-checkLength();
